refactor: reduce session snapshot diff
