+++ packages/control-plane/src/session/event-stream.ts
@@ -1,13 +1,12 @@
 import type { ClientMessage, ServerMessage } from "../types";
-import {
-  sandboxEventSchema,
-  type EventResponse,
-  type ListEventsResponse,
-} from "@open-inspect/shared/types/sandbox-events";
+import type { EventResponse, ListEventsResponse } from "@open-inspect/shared/types/sandbox-events";
 import { encodeEventTimelineCursor, type EventListCursor } from "./event-cursor";
 import type { EventRow } from "./types";
 import type { SessionRepository } from "./repository";
-import type { SessionTimelineEvent } from "@open-inspect/shared/types/server-messages";
+import {
+  sessionTimelineEventSchema,
+  type SessionTimelineEvent,
+} from "@open-inspect/shared/types/server-messages";
 
 const DEFAULT_REPLAY_LIMIT = 500;
 const DEFAULT_HISTORY_LIMIT = 200;
@@ -93,15 +92,13 @@ export class SessionEventStream {
 function parseSessionTimelineEvents(rows: EventRow[]): SessionTimelineEvent[] {
   const events: SessionTimelineEvent[] = [];
   for (const row of rows) {
-    if (!Number.isSafeInteger(row.timeline_sequence) || row.timeline_sequence! < 0) continue;
     try {
-      const event = sandboxEventSchema.safeParse(JSON.parse(row.data));
-      if (!event.success) continue;
-      events.push({
+      const event = sessionTimelineEventSchema.safeParse({
         eventId: row.id,
-        timelineSequence: row.timeline_sequence!,
-        event: event.data,
+        timelineSequence: row.timeline_sequence,
+        event: JSON.parse(row.data),
       });
+      if (event.success) events.push(event.data);
     } catch {
       // A malformed persisted event must not prevent the rest of the timeline from loading.
     }
