+++ packages/web/src/app/(app)/session/error.tsx
