+++ packages/web/src/app/api/sessions/[id]/access/route.ts
@@ -4,12 +4,16 @@ import { controlPlaneUserFetch } from "@/lib/control-plane";
 
 export async function GET(_request: Request, { params }: { params: Promise<{ id: string }> }) {
   const session = await getServerAuthSession();
-  if (!session) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  if (!session?.user) {
+    return NextResponse.json(
+      { error: "Unauthorized" },
+      { status: 401, headers: { "Cache-Control": "private, no-store", Vary: "Cookie" } }
+    );
+  }
 
   const { id } = await params;
   const response = await controlPlaneUserFetch(`/sessions/${encodeURIComponent(id)}/access`, {
     cache: "no-store",
-    headers: { Accept: "application/json" },
   });
   return new Response(response.body, {
     status: response.status,
