+++ packages/web/src/lib/session-socket/reducer.test.ts
@@ -177,17 +177,6 @@ describe("sessionSocketReducer", () => {
         replay: {
           events: [
             {
-              eventId: "event-2",
-              timelineSequence: 2,
-              event: {
-                type: "execution_complete",
-                messageId: "msg-1",
-                success: true,
-                sandboxId: "sb-1",
-                timestamp: 2,
-              },
-            },
-            {
               eventId: "event-1",
               timelineSequence: 1,
               event: {
@@ -196,6 +185,17 @@ describe("sessionSocketReducer", () => {
                 messageId: "msg-1",
                 sandboxId: "sb-1",
                 timestamp: 1,
+              },
+            },
+            {
+              eventId: "event-2",
+              timelineSequence: 2,
+              event: {
+                type: "execution_complete",
+                messageId: "msg-1",
+                success: true,
+                sandboxId: "sb-1",
+                timestamp: 2,
               },
             },
           ],
@@ -232,32 +232,6 @@ describe("sessionSocketReducer", () => {
       });
       expect(state.sessionState?.isProcessing).toBe(true);
       expect(state.sessionState?.totalCost).toBe(1.25);
-    });
-
-    it("hydrates stable replay envelopes", () => {
-      const state = subscribedState({
-        replay: {
-          events: [
-            {
-              eventId: "event-1",
-              timelineSequence: 1,
-              event: {
-                type: "git_sync",
-                status: "completed",
-                sandboxId: "sb-1",
-                timestamp: 1,
-              },
-            },
-          ],
-          hasMore: false,
-          cursor: null,
-        },
-      });
-
-      expect(state.timelineEvents).toEqual([
-        expect.objectContaining({ eventId: "event-1", timelineSequence: 1 }),
-      ]);
-      expect(state.events[0]).toEqual(expect.objectContaining({ status: "completed" }));
     });
   });
 
@@ -360,55 +334,6 @@ describe("sessionSocketReducer", () => {
       expect(state.hasMoreHistory).toBe(false);
       expect(state.cursor).toBeNull();
       expect(state.events.map((event) => event.timestamp)).toEqual([5, 10, 11]);
-    });
-
-    it("collapses token snapshots within a history page", () => {
-      const state = reduce(
-        subscribedState(),
-        serverMessage({
-          type: "history_page",
-          items: [
-            {
-              eventId: "token-1",
-              timelineSequence: 1,
-              event: {
-                type: "token",
-                content: "Partial",
-                messageId: "msg-1",
-                sandboxId: "sb-1",
-                timestamp: 1,
-              },
-            },
-            {
-              eventId: "token-2",
-              timelineSequence: 2,
-              event: {
-                type: "token",
-                content: "Final",
-                messageId: "msg-1",
-                sandboxId: "sb-1",
-                timestamp: 2,
-              },
-            },
-            {
-              eventId: "complete-1",
-              timelineSequence: 3,
-              event: {
-                type: "execution_complete",
-                messageId: "msg-1",
-                success: true,
-                sandboxId: "sb-1",
-                timestamp: 3,
-              },
-            },
-          ],
-          hasMore: false,
-          cursor: null,
-        })
-      );
-
-      expect(state.events.map((event) => event.type)).toEqual(["token", "execution_complete"]);
-      expect(state.events[0]).toEqual(expect.objectContaining({ content: "Final" }));
     });
 
     it("clears a stuck loadingHistory when a new subscribed snapshot arrives", () => {
